backend/requirements.txt
fastapi
uvicorn
openai
httpx
beautifulsoup4
sse-starlette
pydantic
python-dotenv
pytest
pytest-asyncio
requests 